Datasciene_Auburn_workshop/requirements.txt
streamlit
langchain
transformers
openai
python-dotenv